pyproject.toml
[project]
name = "arxiv-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "openai>=1.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
